Create TapeMathSlope.java
This version of tape math finds the slope of a line running down the middle of the rectangle. It turns the robot an amount inversely until the slope no longer exist or in other words the line is vertical and lined up. The while loop condition can be changed to a dead zone instead of exactly zero.